Fix integer width.

+++ include/Core.h
@@ -14,8 +14,8 @@ namespace rat {
 	using U64 = uint64_t;
 
 	using I8 = int64_t;
-	using I16 = int64_t;
-	using I32 = int64_t;
+	using I16 = int16_t;
+	using I32 = int32_t;
 	using I64 = int64_t;
 
 	using B32 = uint32_t;
